+++ extensions/cart_module/config/cart_module.php
@@ -19,7 +19,7 @@
 */
 $config['extension_meta'] = array(
 	'name'        => 'cart_module',
-	'version'     => '1.0',
+	'version'     => '2.0',
 	'type'        => 'module',
 	'title'       => 'Cart',
 	'author'      => 'SamPoyigi',