+++ main/controllers/Checkout.php
@@ -156,21 +156,18 @@ class Checkout extends Main_Controller {
             );
         }
 
+        $data['order_totals'] = array();
         $order_totals = $this->Orders_model->getOrderTotals($order_info['order_id']);
         if ($order_totals) {
-            $data['order_totals'] = array();
-            foreach (array('cart_total', 'coupon', 'delivery', 'taxes') as $key) {
-                foreach ($order_totals as $total) {
-                    if ($key === $total['code']) {
-                        if ($order_type === 'collection' AND $total['code'] === 'delivery') continue;
+            foreach ($order_totals as $total) {
+                if ($order_type === 'collection' AND $total['code'] === 'delivery') continue;
 
-                        $data['order_totals'][] = array(
-                            'code'  => $total['code'],
-                            'title' => $total['title'],
-                            'value' => $this->currency->format($total['value'])
-                        );
-                    }
-                }
+                $data['order_totals'][] = array(
+                    'code'  => $total['code'],
+                    'title' => htmlspecialchars_decode($total['title']),
+                    'value' => $this->currency->format($total['value']),
+                    'priority' => $total['priority'],
+                );
             }
         }